@@ -44,7 +44,11 @@
     "@rollup/plugin-json": "4.1.0",
     "@rollup/plugin-node-resolve": "13.0.6",
     "@rollup/plugin-replace": "3.0.0",
+    "@vue/compiler-sfc": "^3.2.21",
     "@vue/test-utils": "1.2.2",
+    "@vuepress/plugin-active-header-links": "1.8.2",
+    "@vuepress/plugin-google-analytics": "1.8.2",
+    "axios": "0.24.0",
     "babel-core": "6.26.3",
     "babel-jest": "27.3.1",
     "babel-preset-env": "1.7.0",
@@ -61,25 +65,25 @@
     "eslint-plugin-vue": "8.0.3",
     "jest": "27.3.1",
     "jest-serializer-vue": "2.0.2",
+    "less": "4.1.2",
+    "less-loader": "7.3.0",
     "lodash": "4.17.21",
     "prettier": "2.4.1",
     "rollup": "2.59.0",
     "rollup-plugin-filesize": "9.1.1",
     "rollup-plugin-strip-code": "0.2.7",
     "rollup-plugin-uglify": "6.0.4",
-    "rollup-plugin-vue": "5.1.9",
+    "rollup-plugin-vue": "6.0.0",
     "sinon": "12.0.1",
-    "vue": "2.6.14",
+    "vue": "^3",
     "vue-jest": "3.0.7",
     "vue-template-compiler": "2.6.14",
-    "@vuepress/plugin-active-header-links": "1.8.2",
-    "@vuepress/plugin-google-analytics": "1.8.2",
-    "less": "4.1.2",
-    "less-loader": "7.3.0",
     "vuepress": "1.8.2",
     "vuepress-plugin-redirect": "1.2.5",
-    "vuepress-plugin-sitemap": "2.3.1",
-    "axios": "0.24.0"
+    "vuepress-plugin-sitemap": "2.3.1"
+  },
+  "peerDependencies": {
+    "vue": "^3"
   },
   "jest": {
     "setupFilesAfterEnv": [
