@@ -3,7 +3,7 @@
     <!-- dropdown with append-to-body -->
     <dropdown append-to-body>
       <btn class="dropdown-toggle">Dropdown <span class="caret"></span></btn>
-      <template slot="dropdown">
+      <template #dropdown>
         <li><a role="button">Action</a></li>
         <li><a role="button">Another action</a></li>
         <li><a role="button">Something else here</a></li>
@@ -14,7 +14,7 @@
     <!-- dropdown with append-to-body + menu-right -->
     <dropdown append-to-body menu-right>
       <btn class="dropdown-toggle">Menu-Right <span class="caret"></span></btn>
-      <template slot="dropdown">
+      <template #dropdown>
         <li><a role="button">Action</a></li>
         <li><a role="button">Another action</a></li>
         <li><a role="button">Something else here</a></li>
@@ -25,7 +25,7 @@
     <!-- dropdown with append-to-body + dropup -->
     <dropdown append-to-body dropup>
       <btn class="dropdown-toggle">Dropup <span class="caret"></span></btn>
-      <template slot="dropdown">
+      <template #dropdown>
         <li><a role="button">Action</a></li>
         <li><a role="button">Another action</a></li>
         <li><a role="button">Something else here</a></li>
