@@ -2,7 +2,7 @@
   <div class="uiv">
     <dropdown dropup>
       <btn class="dropdown-toggle">Dropup <span class="caret"></span></btn>
-      <template slot="dropdown">
+      <template #dropdown>
         <li><a role="button">Action</a></li>
         <li><a role="button">Another action</a></li>
         <li><a role="button">Something else here</a></li>
