@@ -4,7 +4,7 @@
       <btn type="primary" class="dropdown-toggle"
         >Dropdown <span class="caret"></span
       ></btn>
-      <template slot="dropdown">
+      <template #dropdown>
         <li><a role="button">Action</a></li>
         <li><a role="button">Another action</a></li>
         <li><a role="button">Something else here</a></li>
@@ -15,7 +15,7 @@
     <dropdown>
       <btn type="info">Split Button</btn>
       <btn type="info" class="dropdown-toggle"><span class="caret"></span></btn>
-      <template slot="dropdown">
+      <template #dropdown>
         <li><a role="button">Action</a></li>
         <li><a role="button">Another action</a></li>
         <li><a role="button">Something else here</a></li>
