@@ -1,14 +1,16 @@
 <template>
   <div class="uiv">
     <navbar>
-      <a slot="brand" class="navbar-brand" role="button">Brand</a>
-      <template slot="collapse">
+      <template #brand>
+        <a class="navbar-brand" role="button">Brand</a>
+      </template>
+      <template #collapse>
         <navbar-nav>
           <dropdown tag="li">
             <a class="dropdown-toggle" role="button"
               >Dropdown 1<span class="caret"></span
             ></a>
-            <template slot="dropdown">
+            <template #dropdown>
               <li><a role="button">Action</a></li>
               <li><a role="button">Another action</a></li>
               <li><a role="button">Something else here</a></li>
@@ -20,7 +22,7 @@
             <a class="dropdown-toggle" role="button"
               >Dropdown 2<span class="caret"></span
             ></a>
-            <template slot="dropdown">
+            <template #dropdown>
               <li><a role="button">Action</a></li>
               <li><a role="button">Another action</a></li>
               <li><a role="button">Something else here</a></li>
@@ -34,7 +36,7 @@
             <a class="dropdown-toggle" role="button"
               >Dropdown 3<span class="caret"></span
             ></a>
-            <template slot="dropdown">
+            <template #dropdown>
               <li><a role="button">Action</a></li>
               <li><a role="button">Another action</a></li>
               <li><a role="button">Something else here</a></li>
