@@ -3,9 +3,9 @@
     <btn type="primary" @click="open1 = true">HTML Title</btn>
     <btn type="primary" @click="open2 = true">No Header</btn>
     <modal v-model="open1">
-      <span slot="title"
-        ><i class="glyphicon glyphicon-heart"></i> Modal Title</span
-      >
+      <template #title>
+        <span><i class="glyphicon glyphicon-heart"></i> Modal Title</span>
+      </template>
       <p>This is a modal with HTML title.</p>
     </modal>
     <modal v-model="open2" :header="false">
