@@ -1,8 +1,10 @@
 <template>
   <div class="uiv">
     <navbar>
-      <a slot="brand" class="navbar-brand" href="#">Brand</a>
-      <template slot="collapse">
+      <template #brand>
+        <a class="navbar-brand" href="#">Brand</a>
+      </template>
+      <template #collapse>
         <navbar-nav>
           <li class="active">
             <a role="button">Link <span class="sr-only">(current)</span></a>
@@ -21,7 +23,7 @@
             <a class="dropdown-toggle" role="button"
               >Dropdown <span class="caret"></span
             ></a>
-            <template slot="dropdown">
+            <template #dropdown>
               <li><a role="button">Action</a></li>
               <li><a role="button">Another action</a></li>
               <li><a role="button">Something else here</a></li>
