@@ -1,8 +1,10 @@
 <template>
   <div class="uiv">
     <navbar>
-      <a slot="brand" class="navbar-brand" href="#">Brand</a>
-      <template slot="collapse">
+      <template #brand>
+        <a class="navbar-brand" href="#">Brand</a>
+      </template>
+      <template #collapse>
         <navbar-text right
           >Signed in as <a href="#" class="navbar-link">wxsm</a></navbar-text
         >
