@@ -2,7 +2,7 @@
   <div class="uiv">
     <btn id="btn" type="primary">Popover</btn>
     <popover title="Title" target="#btn">
-      <template slot="popover">
+      <template #popover>
         <h1>Hello world!</h1>
       </template>
     </popover>
