@@ -7,15 +7,20 @@
       <tab title="Profile">
         <p>Profile tab.</p>
       </tab>
-      <form slot="nav-right">
-        <select class="form-control" style="display: inline-block; width: auto">
-          <option>option1</option>
-          <option>option2</option>
-          <option>option3</option>
-          <option>option4</option>
-        </select>
-        <btn type="success">Button</btn>
-      </form>
+      <template #nav-right>
+        <form>
+          <select
+            class="form-control"
+            style="display: inline-block; width: auto"
+          >
+            <option>option1</option>
+            <option>option2</option>
+            <option>option3</option>
+            <option>option4</option>
+          </select>
+          <btn type="success">Button</btn>
+        </form>
+      </template>
     </tabs>
   </div>
 </template>
